updates
merged:

  flora: f85d04192d468a613d5a66df9ab4f9f0216b9eba
  faustoo: 12e3ad44eb6d591b877ffaa3d0625c49401c882f
  fusion809: d548793838004e9535aa584c4ee886e831d2c90e
  deadbeef: 53e0a41db6e7193bd83ef117f8055457f0990cb7
  gentoo-staging: 9292583023a6610905c79a4707db3ac6c63b3a31
  kit-fixups: 27a55db3f6c4399661266975896c28ff3613330c

+++ x11-misc/polybar/polybar-3.4.0.ebuild
@@ -1,24 +1,29 @@
-# Copyright 1999-2018 Gentoo Foundation
+# Copyright 1999-2019 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
 EAPI=6
+
 PYTHON_COMPAT=( python2_7 )
-
 inherit cmake-utils python-single-r1
-
-XPP_VERSION="1.4.0"
-I3IPCPP_VERSION="0.7.1"
 
 DESCRIPTION="A fast and easy-to-use tool for creating status bars"
 HOMEPAGE="https://github.com/jaagr/polybar"
-SRC_URI="https://github.com/jaagr/${PN}/archive/${PV}.tar.gz -> ${P}.tar.gz
-	https://github.com/jaagr/xpp/archive/${XPP_VERSION}.tar.gz -> xpp-${XPP_VERSION}.tar.gz
-	https://github.com/jaagr/i3ipcpp/archive/v${I3IPCPP_VERSION}.tar.gz -> i3ipcpp-${I3IPCPP_VERSION}.tar.gz
-"
+
+if [[ ${PV} != *9999* ]]; then
+	XPP_VERSION="1.4.0"
+	I3IPCPP_VERSION="0.7.1"
+	SRC_URI="https://github.com/jaagr/${PN}/archive/${PV}.tar.gz -> ${P}.tar.gz
+		https://github.com/jaagr/xpp/archive/${XPP_VERSION}.tar.gz -> xpp-${XPP_VERSION}.tar.gz
+		https://github.com/jaagr/i3ipcpp/archive/v${I3IPCPP_VERSION}.tar.gz -> i3ipcpp-${I3IPCPP_VERSION}.tar.gz"
+	KEYWORDS="~amd64 ~x86"
+else
+	inherit git-r3
+	EGIT_REPO_URI="https://github.com/jaagr/${PN}.git"
+	EGIT_CLONE_TYPE="shallow"
+fi
 
 LICENSE="MIT"
 SLOT="0"
-KEYWORDS="amd64 x86"
 
 IUSE="alsa curl i3wm ipc mpd network pulseaudio"
 REQUIRED_USE="${PYTHON_REQUIRED_USE}"
@@ -31,6 +36,7 @@ DEPEND="
 	x11-libs/xcb-util-image
 	x11-libs/xcb-util-wm
 	x11-libs/xcb-util-xrm
+	dev-python/sphinx
 	alsa? ( media-libs/alsa-lib )
 	curl? ( net-misc/curl )
 	i3wm? ( dev-libs/jsoncpp )
@@ -44,13 +50,18 @@ RDEPEND="${DEPEND}"
 src_prepare() {
 	cmake-utils_src_prepare
 
-	rmdir "${S}"/lib/xpp || die
-	mv "${WORKDIR}"/xpp-$XPP_VERSION "${S}"/lib/xpp || die
+	if [[ ${PV} != *9999* ]]; then
+		rmdir "${S}"/lib/xpp || die
+		mv "${WORKDIR}"/xpp-$XPP_VERSION "${S}"/lib/xpp || die
 
-	rmdir "${S}"/lib/i3ipcpp || die
-	mv "${WORKDIR}"/i3ipcpp-$I3IPCPP_VERSION "${S}"/lib/i3ipcpp || die
+		rmdir "${S}"/lib/i3ipcpp || die
+		mv "${WORKDIR}"/i3ipcpp-$I3IPCPP_VERSION "${S}"/lib/i3ipcpp || die
 
-	sed -i "s/.*cpp_error,.*/&\n\t  'eventstruct'   : lambda x, y: None,/" lib/xpp/generators/cpp_client.py || die "sed failed"
+		sed -i "s/.*cpp_error,.*/&\n\t  'eventstruct'   : lambda x, y: None,/" lib/xpp/generators/cpp_client.py || die "sed failed"
+	fi
+	
+	# patch sandbox violation checking if fonts are available
+	epatch "${FILESDIR}/query_fonts_sandbox.patch"
 }
 
 src_configure() {
@@ -63,5 +74,6 @@ src_configure() {
 		-DENABLE_NETWORK="$(usex network)"
 		-DENABLE_PULSEAUDIO="$(usex pulseaudio)"
 	)
+
 	cmake-utils_src_configure
 }
